NYC-Bootstrap/requirements.txt
# S.P.E.C. NYC - Requirements
# Core ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0
optuna>=3.4.0
mlflow>=2.8.0

# Data Validation
pandera>=0.17.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Spatial
h3>=3.7.0
geopandas>=0.14.0

# Dashboard (V1-V2)
streamlit>=1.28.0
plotly>=5.18.0
folium>=0.15.0
streamlit-folium>=0.15.0

# API (V3)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# AI
openai>=1.3.0
chromadb>=0.4.0

# PDF Generation (V3)
weasyprint>=60.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0
requests>=2.31.0
